# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  accountabstractiondto.HandleOpsDTO:
    description: HandleOps request payload with private key, wallet address, target
      address, data, and nonce
    properties:
      data:
        example: 0x...
        type: string
      details:
        $ref: '#/definitions/response.Details'
      nonce:
        example: "0"
        type: string
      priv_key:
        example: 0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef
        type: string
      target:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      wallet:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
    required:
    - data
    - details
    - nonce
    - priv_key
    - target
    - wallet
    type: object
  accountabstractiondto.UserOpDTO:
    description: UserOp object containing target address, value, data, client block
      number, user nonce, and signature
    properties:
      client_block_number:
        example: "12345"
        type: string
      data:
        example: 0x...
        type: string
      signature:
        example: 0x...
        type: string
      target:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
      user_nonce:
        example: "0"
        type: string
      value:
        example: "0"
        type: string
    required:
    - client_block_number
    - data
    - signature
    - target
    - user_nonce
    - value
    type: object
  accountabstractiondto.ValidateUserOpsDTO:
    description: Validation payload with wallet address and UserOp object
    properties:
      input:
        $ref: '#/definitions/accountabstractiondto.UserOpDTO'
      wallet:
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        type: string
    required:
    - input
    - wallet
    type: object
...
  response.BlockchainResponse:
    properties:
      errors:
        example: No Error Message
        type: string
      returned: {}
      success:
        example: true
        type: boolean
    type: object
  response.Details:
    properties:
      service:
        example: 0
        type: integer
      tx_type:
        example: 10
        type: integer
    type: object
  response.Meta:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: DID Backend API
  title: DID API
  version: "1.0"
paths:
...
  /account-abstraction/handle-ops:
    post:
      consumes:
      - application/json
      description: |-
        Executes user operations through the Account Abstraction entry point. This endpoint processes user operations by signing and submitting them to the blockchain via the entry point contract.

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true for HTTP 200
        - `data` (object): Map where each key is a transaction hash and value contains transaction result
        - `meta` (object): Contains timestamp and API version

        ## Transaction Result Object
        Each transaction hash in the data map contains:
        - `success` (boolean): Indicates if the transaction was successful on the blockchain
        - `errors` (string): Error message if failed, or "No Error Message" if successful
        - `returned` (object|null): CreateDIDEvent object containing DID creation details, or null if transaction failed or event decoding failed

        ## Common Error Scenarios
        ### Transaction Errors (success: false in transaction result):
        - "execution reverted" - Transaction reverted on the blockchain
        - "nonce too low" - Transaction nonce conflict
        - "transaction underpriced" - Gas price below network minimum
        - "invalid signature" - Transaction signing failed
        - "insufficient funds" - Relayer lacks funds for gas fees

        ## Important Notes
        - HTTP 200 status does NOT guarantee operation success
        - Always check the `success` field within each transaction result in the data map
        - Multiple transaction hashes may be returned for batch operations
        - The PrivKey is used by the relayer to sign and submit the transaction
      parameters:
      - description: HandleOps payload with private key, wallet address, target address,
          data, and nonce
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accountabstractiondto.HandleOpsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Transaction(s) processed successfully - check individual transaction
            results
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    $ref: '#/definitions/response.BlockchainResponse'
                  type: object
              type: object
        "400":
          description: Invalid request body - malformed JSON, missing required fields,
            or invalid address format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - handle ops use case failed, network
            connectivity issues, RPC node errors, or transaction submission failure
          schema:
            $ref: '#/definitions/response.Response'
      summary: Handle User Operations
      tags:
      - AA POST
  /account-abstraction/is-valid-wallet:
    get:
      consumes:
      - application/json
      description: |-
        Checks whether a given wallet address is a valid Account Abstraction smart wallet.

        This endpoint verifies:
        - The wallet contract exists on-chain
        - The wallet conforms to the expected smart wallet implementation
        - The wallet can be used with the Account Abstraction entry point

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true if request is processed correctly
        - `data` (object): Validation result returned from the blockchain
        - `meta` (object): Contains timestamp and API version

        ## Important Notes
        - HTTP 200 does NOT guarantee the wallet is valid
        - Always check the `data` field for validation result
        - This endpoint does NOT modify blockchain state (read-only)
      parameters:
      - description: Smart wallet address to validate
        example: 0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb
        in: query
        name: wallet
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wallet validation result
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.BlockchainResponse'
              type: object
        "400":
          description: Invalid request or invalid wallet address format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - blockchain call failed or RPC error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Validate Smart Wallet
      tags:
      - AA GET
  /account-abstraction/validate-userop:
    post:
      consumes:
      - application/json
      description: |-
        Validates a user operation for an Account Abstraction wallet. This endpoint checks if a user operation is valid before execution, ensuring the operation can be processed by the wallet contract.

        ## Response Structure
        Success responses (HTTP 200) contain:
        - `success` (boolean): Always true for HTTP 200
        - `data` (string): Validation result message ("UserOp Valid")
        - `meta` (object): Contains timestamp and API version

        ## UserOp Structure
        The UserOp object contains:
        - `Target` (string): The target contract address for the operation
        - `Value` (string): The amount of ETH to send (in wei, as hex string)
        - `Data` (string): The encoded function call data (hex string)
        - `ClientBlockNumber` (string): The client block number (as hex string)
        - `UserNonce` (string): The user operation nonce (as hex string)
        - `Signature` (string): The signature for the user operation (hex string)

        ## Common Error Scenarios
        ### Validation Errors:
        - "invalid user operation" - UserOp structure is invalid
        - "signature verification failed" - Signature does not match the operation
        - "nonce mismatch" - UserNonce does not match expected value
        - "insufficient balance" - Wallet lacks sufficient funds
        - "invalid target address" - Target contract address is invalid

        ## Important Notes
        - This endpoint only validates the operation, it does not execute it
        - The wallet implementation address is automatically updated before validation
        - Validation ensures the operation can be safely executed
      parameters:
      - description: Validation payload with wallet address and UserOp object
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accountabstractiondto.ValidateUserOpsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: User operation is valid
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Invalid request body - malformed JSON, missing required fields,
            or invalid address format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error - validation failed, network connectivity
            issues, or RPC node errors
          schema:
            $ref: '#/definitions/response.Response'
      summary: Validate User Operation
      tags:
      - AA POST
...
swagger: "2.0"