# internal/delivery/http/handler/accountabstraction/entrypoint.go
package accountabstractionhandler

import (
	"github.com/gofiber/fiber/v2"
	"github.com/meQlause/go-be-did/internal/config"
	"github.com/meQlause/go-be-did/internal/validator"
	"github.com/meQlause/go-be-did/pkg/response"

	didevent "github.com/meQlause/go-be-did/internal/delivery/event/handleops"
	accountabstractiondto "github.com/meQlause/go-be-did/internal/domain/dtos/accountabstraction"
	aasdk "github.com/meQlause/go-be-did/internal/infrastructure/sdk/accountabstraction"
)

// HandleOps godoc
// @Summary      Handle User Operations
// @Description  Executes user operations through the Account Abstraction entry point. This endpoint processes user operations by signing and submitting them to the blockchain via the entry point contract.
// @Description
// @Description  ## Response Structure
// @Description  Success responses (HTTP 200) contain:
// @Description  - `success` (boolean): Always true for HTTP 200
// @Description  - `data` (object): Map where each key is a transaction hash and value contains transaction result
// @Description  - `meta` (object): Contains timestamp and API version
// @Description
// @Description  ## Transaction Result Object
// @Description  Each transaction hash in the data map contains:
// @Description  - `success` (boolean): Indicates if the transaction was successful on the blockchain
// @Description  - `errors` (string): Error message if failed, or "No Error Message" if successful
// @Description  - `returned` (object|null): CreateDIDEvent object containing DID creation details, or null if transaction failed or event decoding failed
// @Description
// @Description  ## Common Error Scenarios
// @Description  ### Transaction Errors (success: false in transaction result):
// @Description  - "execution reverted" - Transaction reverted on the blockchain
// @Description  - "nonce too low" - Transaction nonce conflict
// @Description  - "transaction underpriced" - Gas price below network minimum
// @Description  - "invalid signature" - Transaction signing failed
// @Description  - "insufficient funds" - Relayer lacks funds for gas fees
// @Description
// @Description  ## Important Notes
// @Description  - HTTP 200 status does NOT guarantee operation success
// @Description  - Always check the `success` field within each transaction result in the data map
// @Description  - Multiple transaction hashes may be returned for batch operations
// @Description  - The PrivKey is used by the relayer to sign and submit the transaction
// @Tags         AA POST
// @Accept       json
// @Produce      json
// @Param        request body accountabstractiondto.HandleOpsDTO true "HandleOps payload with private key, wallet address, target address, data, and nonce"
// @Success      200 {object} response.Response{data=map[string]response.BlockchainResponse} "Transaction(s) processed successfully - check individual transaction results"
// @Failure      400 {object} response.Response "Invalid request body - malformed JSON, missing required fields, or invalid address format"
// @Failure      500 {object} response.Response "Internal server error - handle ops use case failed, network connectivity issues, RPC node errors, or transaction submission failure"
// @Router       /account-abstraction/handle-ops [post]
func (ah *AccountAbstractionHandler) HandleOps(c *fiber.Ctx) error {
	var input accountabstractiondto.HandleOpsDTO
	if err := c.BodyParser(&input); err != nil {
		return response.Error(c, fiber.StatusBadRequest, err.Error())
	}

	if err := validator.Validate.Struct(&input); err != nil {
		validationErrors := validator.FormatError(err)
		return response.Error(c, fiber.StatusBadRequest, validationErrors)
	}

	handleOpsInput := input.Into()
	aasdk.ChangeWalletImplementationAddress(handleOpsInput.Wallet, config.Blockchain())

	result, err := ah.uc.HandleOps(c.Context(), handleOpsInput, config.Network())
	if err != nil {
		return response.Error(c, fiber.StatusInternalServerError, err.Error())
	}

	txSuccess, txErrors := config.Blockchain().CheckTxs(c.Context(), result.TxHash)

	resp := make(map[string]response.BlockchainResponse)
	decodeEventFunc, _ := didevent.Registry[input.Details.Service][input.Details.TxType]

	for txHash, ok := range txSuccess {
		eventData, err := decodeEventFunc(c.Context(), txHash)
		if err != nil {
			continue
		}

		resp[txHash.Hex()] = response.BlockchainResponse{
			Success:  ok,
			Errors:   "No Error Message",
			Returned: eventData,
		}
	}

	for txHash, txErr := range txErrors {
		resp[txHash.Hex()] = response.BlockchainResponse{
			Success:  false,
			Errors:   txErr.Error(),
			Returned: nil,
		}
	}

	return response.Success(c, resp)
}

// IsValidWallet godoc
// @Summary      Validate Smart Wallet
// @Description  Checks whether a given wallet address is a valid Account Abstraction smart wallet.
// @Description
// @Description  This endpoint verifies:
// @Description  - The wallet contract exists on-chain
// @Description  - The wallet conforms to the expected smart wallet implementation
// @Description  - The wallet can be used with the Account Abstraction entry point
// @Description
// @Description  ## Response Structure
// @Description  Success responses (HTTP 200) contain:
// @Description  - `success` (boolean): Always true if request is processed correctly
// @Description  - `data` (object): Validation result returned from the blockchain
// @Description  - `meta` (object): Contains timestamp and API version
// @Description
// @Description  ## Important Notes
// @Description  - HTTP 200 does NOT guarantee the wallet is valid
// @Description  - Always check the `data` field for validation result
// @Description  - This endpoint does NOT modify blockchain state (read-only)
// @Tags         AA GET
// @Accept       json
// @Produce      json
// @Param        wallet query string true "Smart wallet address to validate" example(0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb)
// @Success      200 {object} response.Response{data=response.BlockchainResponse} "Wallet validation result"
// @Failure      400 {object} response.Response "Invalid request or invalid wallet address format"
// @Failure      500 {object} response.Response "Internal server error - blockchain call failed or RPC error"
// @Router       /account-abstraction/is-valid-wallet [get]
func (ah *AccountAbstractionHandler) IsValidWallet(c *fiber.Ctx) error {
	var input accountabstractiondto.IsValidWalletDTO
	if err := c.QueryParser(&input); err != nil {
		return response.Error(c, fiber.StatusBadRequest, err.Error())
	}

	if err := validator.Validate.Struct(&input); err != nil {
		validationErrors := validator.FormatError(err)
		return response.Error(c, fiber.StatusBadRequest, validationErrors)
	}

	IsValidWalletInput := input.Into()

	result, err := ah.uc.IsValidWallet(c.Context(), IsValidWalletInput)
	if err != nil {
		return response.Error(c, fiber.StatusInternalServerError, err.Error())
	}

	resp := response.BlockchainResponse{
		Success:  true,
		Errors:   "No Error Message",
		Returned: result,
	}

	return response.Success(c, resp)
}